—— 一切就绪，只差pid

+++ Firmware/HeiGuan_Release_v1.0/Hardware/PWM.h
@@ -10,6 +10,5 @@
 
 /* Public Functions ------------------------------------------------------------------*/
 void Motor_Init();
-void Motor_SetSpeed_All(int8_t speed);
+void Motor_SetSpeed_All(uint16_t speed);
 #endif
-
